ADRP/requirements.txt
asgiref
attrs
boto3
botocore==1.37.5
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==25.1.0
django-storages==1.14.5
djangorestframework==3.15.2
environ
psycopg2-binary
djangorestframework-simplejwt
django-extensions
django-cors-headers
PyJWT
gunicorn
drf-spectacular
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
environ==1.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.10.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
s3transfer==0.11.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
django-filter
djangorestframework-recaptcha
